Broker Based distributed connection
The following project is a AUEB project that grants a successful connection between brokers remotely.
Although it lacks a user friendly and welcoming user interface because its based on the back end side of the connections.
Thanks to that connection the users(brokers) can share text messages images and video files